requirements.txt
altgraph==0.17
beautifulsoup4==4.9.3
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
cloudscraper==1.2.52
cycler==0.10.0
docopt==0.6.2
et-xmlfile==1.0.1
future==0.18.2
google-api-core==1.21.0
google-api-python-client==1.12.5
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
jdcal==1.4.1
kiwisolver==1.3.1
lxml==4.6.1
matplotlib==3.3.3
multitasking==0.0.9
numpy==1.19.5
oauthlib==3.1.0
openpyxl==3.0.6
pandas==1.2.1
pandas-datareader==0.9.0
pefile==2019.4.18
Pillow==8.1.0
pipreqs==0.4.10
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
undetected-chromedriver==2.0.2
uritemplate==3.0.1
urllib3==1.25.9
xlrd==2.0.1
yarg==0.1.9
yfinance==0.1.59
